+++ tensorflow/engine/src/main/java/ai/djl/tensorflow/engine/TfNDArray.java
@@ -1002,7 +1002,7 @@ public class TfNDArray implements NDArray {
 
     /** {@inheritDoc} */
     @Override
-    public NDArray argMax(long axis) {
+    public NDArray argMax(int axis) {
         return null;
     }
 
@@ -1014,7 +1014,7 @@ public class TfNDArray implements NDArray {
 
     /** {@inheritDoc} */
     @Override
-    public NDArray argMin(long axis) {
+    public NDArray argMin(int axis) {
         return null;
     }
